refactor(recovery): extract shared dialog primitives to _shared.tsx
Review m3: `PasswordField`, `StrengthMeter`, `useLiveStrength`, and
the error-coercion helper were copy-pasted between
AccountRecoveryDialog.tsx and ExistingUserRecoveryPrompt.tsx. Extract
into `_shared.tsx` so the signup dialog and the migration nag stay in
lockstep. ~70 lines deleted.
